requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.7.4
pycryptodome==3.20.0
stripe==10.3.0
httpx==0.27.0
python-multipart==0.0.9
